third_party/blink/web_tests/external/wpt/html/capability-delegation/delegation-consumes-activation.https.tentative.html

<!DOCTYPE html>
<!--
   Tentative due to:
     https://github.com/whatwg/html/issues/4008
-->
<title>Capability Delegation: Consumes User Activation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/utils.js"></script>

<div>
  Test that capability delegation consumes transient user activation.

  https://wicg.github.io/capability-delegation/spec.html
</div>

<iframe allow="payment" width="300px" height="50px" src="about:blank"></iframe>

<script>
  promise_test(async () => {
      assert_false(navigator.userActivation.isActive);
      await test_driver.bless();
      assert_true(navigator.userActivation.isActive);
      frames[0].postMessage({"type": "none"}, {targetOrigin: location.origin, delegate: ""});
      assert_true(navigator.userActivation.isActive);
      frames[0].postMessage({"type": "none"}, {targetOrigin: location.origin, delegate: "payment"});
      assert_false(navigator.userActivation.isActive);
  }, `capability delegation consumes transient user activation`);
</script>
